Close #61
