+++ .idea/misc.xml
@@ -4,7 +4,7 @@
     <resource url="http://schemas.android.com/" location="$PROJECT_DIR$" />
     <resource url="http://schemas.android.com/apk/res/android" location="$PROJECT_DIR$" />
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_1_8" default="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
