+++ app/src/main/java/com/example/myapplication/PongActivity.java
@@ -1,4 +1,30 @@
 package com.example.myapplication;
 
-public class PongActivity {
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.view.WindowManager;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import pong.GameActivity;
+
+public class PongActivity extends AppCompatActivity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN);
+        setContentView(R.layout.activity_pong);
+
+
+
+        findViewById(R.id.play).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                startActivity(new Intent(PongActivity.this, GameActivity.class));
+                //setContentView(R.layout.pong_game);
+            }
+        });
+    }
 }
